Polynomial: int coefficients instead of ll, 1.3-1.5x faster
The std_abs copies of this chapter really were faster, and the reason
is not the algorithm -- it is that they keep coefficients in int while
we kept them in ll.  P < 2^31, so int holds every coefficient; going
from 8 bytes to 4 halves the memory traffic through the transform,
which is what an NTT is bound by.

NTT<MAXN, int P, int RT> now holds int w[MAXN] and transforms int *a;
Poly derives from vector<int>.  Every product is cast one at a time,
(ll)a * b % P.  The two places where a sum of three residues could
have overflowed int (PowerProj's b[n2 + i]) take the cast as well.

Library Checker, -j 1, before -> after:

  inv                0.617s -> 0.467s
  exp                2.221s -> 1.520s
  log                0.869s -> 0.667s
  pow                3.174s -> 2.377s
  sqrt               1.570s -> 1.118s
  taylor_shift       0.368s -> 0.267s
  sampling_points    0.868s -> 0.667s
  convolution_mod    0.371s -> 0.326s

That beats the std_abs versions on the same data (their exp 1.618s,
sqrt 1.618s), and unlike them Pow terminates: std_abs's PolyPow passes
a 1e18 k into Pow(int, int), and the truncated negative b never
reaches 0 under b >>= 1, so it hangs on all four 5e5 cases.

48 AC / 58 rows, 0 WA, unchanged.  kth_term_of_linearly_recurrent_
sequence improves from 142% to 115% of TL but still TLEs; that one is
DivMod recomputing Inv(C), not the coefficient type.

API change for callers: Eval/Interpolate/SampleShift/LinearRecursion/
PowerProj take and return vector<int> now, and LinearRecursion returns
int.  Pow still takes ll k.

+++ codebook/7_Polynomial/Number_Theory_Transform.cpp
@@ -2,17 +2,18 @@
 // 7*17*(2^23)+1, 998244353, 3
 // 1255*(2^20)+1, 1315962881, 3
 // 51*(2^25)+1, 1711276033, 29
-template <int MAXN, ll P, ll RT> // MAXN must be 2^k
+template <int MAXN, int P, int RT> // MAXN must be 2^k
 struct NTT {
-  ll w[MAXN];
-  ll mpow(ll a, ll n);
-  ll minv(ll a) { return mpow(a, P - 2); }
+  int w[MAXN];
+  int mpow(int a, ll n);
+  int minv(int a) { return mpow(a, P - 2); }
   NTT() {
-    ll dw = mpow(RT, (P - 1) / MAXN);
+    int dw = mpow(RT, (P - 1) / MAXN);
     w[0] = 1;
-    FOR (i, 1, MAXN - 1) w[i] = w[i - 1] * dw % P;
+    FOR (i, 1, MAXN - 1)
+      w[i] = (ll)w[i - 1] * dw % P;
   }
-  void bitrev(ll *a, int n) {
+  void bitrev(int *a, int n) {
     int i = 0;
     FOR (j, 1, n - 2) {
       for (int k = n >> 1; (i ^= k) < k; k >>= 1);
@@ -20,14 +21,14 @@ struct NTT {
     }
   }
   void operator()(
-    ll *a, int n, bool inv = false) { // 0 <= a[i] < P
+    int *a, int n, bool inv = false) { // 0 <= a[i] < P
     bitrev(a, n);
     for (int L = 2; L <= n; L <<= 1) {
       int dx = MAXN / L, dl = L >> 1;
       for (int i = 0; i < n; i += L) {
         for (int j = i, x = 0; j < i + dl;
           j++, x += dx) {
-          ll tmp = a[j + dl] * w[x] % P;
+          int tmp = (ll)a[j + dl] * w[x] % P;
           if ((a[j + dl] = a[j] - tmp) < 0)
             a[j + dl] += P;
           if ((a[j] += tmp) >= P) a[j] -= P;
@@ -36,8 +37,8 @@ struct NTT {
     }
     if (inv) {
       reverse(a + 1, a + n);
-      ll invn = minv(n);
-      FOR (i, 0, n - 1) a[i] = a[i] * invn % P;
+      int invn = minv(n);
+      FOR (i, 0, n - 1) a[i] = (ll)a[i] * invn % P;
     }
   }
 };
